+++ src/risktools/charts.py
@@ -8,6 +8,9 @@ from plotly.colors import DEFAULT_PLOTLY_COLORS
 from math import ceil
 from ._morningstar import *
 from statsmodels.tsa.seasonal import STL
+from .pa import *
+
+from .main_functions import get_eia_df
 
 
 def chart_zscore(df, freq=None, output="zscore", chart="seasons", **kwargs):